:sparkles: Changed main activity to `LoadingActivity`
