+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/LoadingActivity.kt
@@ -1,17 +1,23 @@
 package com.arnyminerz.escalaralcoiaicomtat.activity
 
 import android.content.Intent
+import android.os.Build
 import android.os.Bundle
 import androidx.annotation.UiThread
 import com.arnyminerz.escalaralcoiaicomtat.R
 import com.arnyminerz.escalaralcoiaicomtat.activity.model.NetworkChangeListenerActivity
 import com.arnyminerz.escalaralcoiaicomtat.appNetworkState
+import com.arnyminerz.escalaralcoiaicomtat.data.IntroShowReason
 import com.arnyminerz.escalaralcoiaicomtat.data.climb.data.loadAreasFromCache
 import com.arnyminerz.escalaralcoiaicomtat.databinding.ActivityLoadingBinding
 import com.arnyminerz.escalaralcoiaicomtat.exception.NoInternetAccessException
 import com.arnyminerz.escalaralcoiaicomtat.generic.runAsync
 import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
+import com.arnyminerz.escalaralcoiaicomtat.notification.createNotificationChannels
 import com.arnyminerz.escalaralcoiaicomtat.view.visibility
+import io.sentry.SentryLevel
+import io.sentry.android.core.SentryAndroid
+import io.sentry.android.timber.SentryTimberIntegration
 import timber.log.Timber
 
 class LoadingActivity : NetworkChangeListenerActivity() {
@@ -23,6 +29,29 @@ class LoadingActivity : NetworkChangeListenerActivity() {
         binding = ActivityLoadingBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
+        Timber.plant(Timber.DebugTree())
+        Timber.v("Planted Timber.")
+
+        Timber.v("Instantiating Sentry")
+        SentryAndroid.init(this) { options ->
+            options.addIntegration(
+                SentryTimberIntegration(SentryLevel.ERROR, SentryLevel.INFO)
+            )
+        }
+
+        val showIntro = IntroActivity.shouldShow(this)
+        if (showIntro != IntroShowReason.OK) {
+            Timber.w("  Showing intro! Reason: ${showIntro.msg}")
+            finish()
+            startActivity(Intent(this, IntroActivity::class.java))
+            return
+        } else
+            Timber.v("  Won't show intro.")
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O)
+            createNotificationChannels()
+
+        Timber.v("Finished preparing App...")
         load()
     }
